Specify worker resources

Specify resources for worker processes. If given, the values here will override those in didehpc_config(). See vignette("workers") for more details.

worker_resource(
  template = NULL,
  cores = NULL,
  wholenode = NULL,
  parallel = NULL
)

Arguments

template
A job template. On fi--dideclusthn this can be "GeneralNodes" or "8Core", while on "fi--didemrchnb" this can be "GeneralNodes", "12Core", "16Core", "12and16Core", "20Core", "24Core" or "32Core". See the main cluster documentation if you tweak these parameters, as you may not have permission to use all templates (and if you use one that you don't have permission for the job will fail). For training purposes there is also a "Training" template, but you will only need to use this when instructed to.

cores
The number of cores to request. This is mostly useful when using the GeneralNodes template. If specified, then we will request this many cores from the windows queuer. If you request too many cores then your task will queue forever! 24 is the largest this should be on fi--dideclusthn and 64 on fi--didemrchnb (assuming you have access to those nodes). If omitted then a single core is selected for the GeneralNodes template or the entire machine for the other templates (unless modified by wholenode).

wholenode
Request the whole node? This will default to TRUE if any template other than GeneralNodes is selected.

parallel
Should we set up the parallel cluster? Normally if more than one core is implied (via the cores argument, by picking a template other than GeneralNodes or by using wholenode) then a parallel cluster will be set up (see Details). If parallel is set to FALSE then this will not occur. This might be useful in cases where you want to manage your own job level parallelism (e.g. using OpenMP) or if you're just after the whole node for the memory).

Value

A list with class worker_resource which can be passed into didehpc_config